@@ -3,7 +3,7 @@
   "version": "0.2.0",
   "license": "GPL-3.0",
   "engines": {
-    "node": ">=12"
+    "node": ">=16"
   },
   "private": true,
   "scripts": {
@@ -90,7 +90,10 @@
   "resolutions": {
     "typescript": "4.8.4",
     "@types/react": "18.0.21",
-    "@types/node": "18.11.4"
+    "@types/node": "18.11.4",
+    "@emotion/react": "11.10.4",
+    "@emotion/cache": "11.10.3",
+    "react-select": "5.6.0"
   },
   "browserslist": [
     "defaults",
