@@ -1,4 +1,5 @@
-import { Center, CenterProps, Spinner } from '@chakra-ui/react';
+import type { CenterProps } from '@chakra-ui/react';
+import { Center, Spinner } from '@chakra-ui/react';
 import React from 'react';
 
 export const LoadingState: React.FC<
