@@ -1,13 +1,13 @@
-export { BrightIdIcon } from './BrightIdIcon';
-export { ChainIcon } from './ChainIcon';
-export { Dashboard } from './Dashboard';
-export { DropDownIcon } from './DropDownIcon';
-export { EthereumIcon } from './EthereumIcon';
-export { GnosisIcon } from './GnosisIcon';
-export { Icon3box } from './Icon3box';
-export { LogOut } from './LogOut';
-export { MeetWithWalletIcon } from './MeetWithWalletIcon';
-export { MenuArrow } from './MenuArrow';
-export { MetaMaskIcon } from './MetaMaskIcon';
-export { PolygonIcon } from './PolygonIcon';
-export { Profile } from './Profile';
+export { BrightIdIcon } from './BrightIdIcon.js';
+export { ChainIcon } from './ChainIcon.js';
+export { Dashboard } from './Dashboard.js';
+export { DropDownIcon } from './DropDownIcon.js';
+export { EthereumIcon } from './EthereumIcon.js';
+export { GnosisIcon } from './GnosisIcon.js';
+export { Icon3box } from './Icon3box.js';
+export { LogOut } from './LogOut.js';
+export { MeetWithWalletIcon } from './MeetWithWalletIcon.js';
+export { MenuArrow } from './MenuArrow.js';
+export { MetaMaskIcon } from './MetaMaskIcon.js';
+export { PolygonIcon } from './PolygonIcon.js';
+export { Profile } from './Profile.js';
